[feat/#4] refactor: domain과 web을 분리
