+++ be/src/main/java/airdnb/be/domain/member/MemberRepository.java
@@ -1,6 +1,6 @@
-package airdnb.be.member;
+package airdnb.be.domain.member;
 
-import airdnb.be.member.entitiy.Member;
+import airdnb.be.domain.member.entitiy.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
