+++ be/src/main/java/airdnb/be/web/member/request/EmailRequest.java
@@ -1,4 +1,4 @@
-package airdnb.be.member.controller.request;
+package airdnb.be.web.member.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
