+++ be/src/main/java/airdnb/be/web/member/request/MemberSaveRequest.java
@@ -1,13 +1,12 @@
-package airdnb.be.member.controller.request;
+package airdnb.be.web.member.request;
 
-import airdnb.be.member.entitiy.Member;
+import airdnb.be.domain.member.entitiy.Member;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 
-
-public record MemberRequest(
+public record MemberSaveRequest(
 
         @NotBlank
         String name,
